Player Stats Display & Buff Overhaul
Created a stat display for the player. Now the player's base stats can be viewed as well as current buff effects and their values: Attack, Defense, Hits, Crit, Pierce, Sharp, & Heavy

Added an animation for the stat display to slide in & slide out when the play taps the sword icon on their HUD.

Reworked how the buff system works since it was unnecessarily complicated & made it way to hard to create this simple stat display.