Polish Started
Fixed the camera focusing on enemy turns, when regen actually works, fixed enemy death animations, stairs waiting for player to step on before changing and gave the chest a second before opening. Also added a button that players can use to jump to the next floor without having to go to the stairs if all the conditions are met.